+++ frontend/.next/static/chunks/webpack.js
@@ -192,7 +192,7 @@
 /******/ 	
 /******/ 	/* webpack/runtime/getFullHash */
 /******/ 	!function() {
-/******/ 		__webpack_require__.h = function() { return "ab9f8bf0b1f569b4"; }
+/******/ 		__webpack_require__.h = function() { return "a96454dd971135d7"; }
 /******/ 	}();
 /******/ 	
 /******/ 	/* webpack/runtime/global */