+++ frontend/.next/types/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,4 @@
-// File: C:\Users\jizar\OneDrive\Documents\DiscordBot\frontend\app\api\auth\[...nextauth]\route.ts
+// File: /mnt/c/Users/jizar/OneDrive/Documents/DiscordBot/frontend/app/api/auth/[...nextauth]/route.ts
 import * as entry from '../../../../../../app/api/auth/[...nextauth]/route.js'
 import type { NextRequest } from 'next/server.js'