+++ frontend/.next/types/app/layout.ts
@@ -1,4 +1,4 @@
-// File: C:\Users\jizar\OneDrive\Documents\DiscordBot\frontend\app\layout.tsx
+// File: /mnt/c/Users/jizar/OneDrive/Documents/DiscordBot/frontend/app/layout.tsx
 import * as entry from '../../../app/layout.js'
 import type { ResolvingMetadata, ResolvingViewport } from 'next/dist/lib/metadata/types/metadata-interface.js'